+++ src/commands/openLocalFileCommand.ts
@@ -21,7 +21,7 @@ import { CICSTree } from "../trees/CICSTree";
 import * as https from "https";
 import { CICSRegionsContainer } from "../trees/CICSRegionsContainer";
 import { CICSLocalFileTreeItem } from "../trees/treeItems/CICSLocalFileTreeItem";
-import { findSelectedNodes } from "../utils/commandUtils";
+import { findSelectedNodes, splitCmciErrorMessage } from "../utils/commandUtils";
 import { CICSCombinedLocalFileTree } from "../trees/CICSCombinedLocalFileTree";
 
 export function getOpenLocalFileCommand(tree: CICSTree, treeview: TreeView<any>) {
@@ -69,23 +69,7 @@ export function getOpenLocalFileCommand(tree: CICSTree, treeview: TreeView<any>)
             // @ts-ignore
             if (error.mMessage) {
               // @ts-ignore
-              const mMessageArr = error.mMessage.split(" ").join("").split("\n");
-              let resp;
-              let resp2;
-              let respAlt;
-              let eibfnAlt;
-              for (const val of mMessageArr) {
-                const values = val.split(":");
-                if (values[0] === "resp"){
-                  resp = values[1];
-                } else if (values[0] === "resp2"){
-                  resp2 = values[1];
-                } else if (values[0] === "resp_alt"){
-                  respAlt = values[1];
-                } else if (values[0] === "eibfn_alt"){
-                  eibfnAlt = values[1];
-                }
-              }
+              const [_, resp2, respAlt, eibfnAlt] = splitCmciErrorMessage(error.mMessage);
               window.showErrorMessage(`Perform OPEN on local file "${allSelectedNodes[parseInt(index)].localFile.file}" failed: EXEC CICS command (${eibfnAlt}) RESP(${respAlt}) RESP2(${resp2})`);
             } else {
               window.showErrorMessage(`Something went wrong when performing an OPEN - ${JSON.stringify(error, Object.getOwnPropertyNames(error)).replace(/(\\n\t|\\n|\\t)/gm," ")}`);
