+++ src/commands/phaseInCommand.ts
@@ -17,7 +17,7 @@ import { CICSTree } from "../trees/CICSTree";
 import * as https from "https";
 import { CICSRegionsContainer } from "../trees/CICSRegionsContainer";
 import { CICSProgramTreeItem } from "../trees/treeItems/CICSProgramTreeItem";
-import { findSelectedNodes } from "../utils/commandUtils";
+import { findSelectedNodes, splitCmciErrorMessage } from "../utils/commandUtils";
 import { CICSCombinedProgramTree } from "../trees/CICSCombinedProgramTree";
 
 /**
@@ -71,23 +71,7 @@ import { CICSCombinedProgramTree } from "../trees/CICSCombinedProgramTree";
             // @ts-ignore
             if (error.mMessage) {
               // @ts-ignore
-              const mMessageArr = error.mMessage.split(" ").join("").split("\n");
-              let resp;
-              let resp2;
-              let respAlt;
-              let eibfnAlt;
-              for (const val of mMessageArr) {
-                const values = val.split(":");
-                if (values[0] === "resp"){
-                  resp = values[1];
-                } else if (values[0] === "resp2"){
-                  resp2 = values[1];
-                } else if (values[0] === "resp_alt"){
-                  respAlt = values[1];
-                } else if (values[0] === "eibfn_alt"){
-                  eibfnAlt = values[1];
-                }
-              }
+              const [_, resp2, respAlt, eibfnAlt] = splitCmciErrorMessage(error.mMessage);
               window.showErrorMessage(`Perform PHASEIN on Program "${allSelectedNodes[parseInt(index)].program.program}" failed: EXEC CICS command (${eibfnAlt}) RESP(${respAlt}) RESP2(${resp2})`);
             } else {
               window.showErrorMessage(`Something went wrong when performing a PHASEIN - ${JSON.stringify(error, Object.getOwnPropertyNames(error)).replace(/(\\n\t|\\n|\\t)/gm," ")}`);
